Make plans produce string messages instead of java objects
Instead of creating a SumoCommand directly,
we create PlanMessages which can generate the SumoCommand.

We are using Gson for serialization.
Gson can't serialize objects generic types,
unless the explictly instanciated type is given
during both serialization and desiralization.
SumoCommand instances contain an instance attribute
of type List<Object>.
Without going into the details of SumoCommand implementation,
it is hard to know how to serialize that.
Further the explicit class of the object being serialized
is required to be know when desializing.

Thus we use PlanMessage and PlanMessageParser
to serialize an deserialize messages.