Fixed bug where agents wouldn't enter environment after route finished.
SUMO automatically removes vehicles that have arrived at their destination. During
an update to this code, the logic for re-entering as soon as that happened was
deactivated. It has been activated again.

Because of a bug, performance degraged over time. This bug has been fixed [1]

Clean up of unused code, and spelling error fixes.

[1] Agents adopted both maintenance goals each time they entered the environment. For
each of these goals, a plan would fire. So if an agent re-entered 30 times,
2 * 30 = 60 plans would fire for that agent in one deliberation cycle, where 2 is enough.